+++ testing/simulated/valid_chain_test.go
@@ -1,0 +1,70 @@
+//go:build simulated
+
+// SPDX-License-Identifier: BUSL-1.1
+//
+// Copyright (C) 2025, Berachain Foundation. All rights reserved.
+// Use of this software is governed by the Business Source License included
+// in the LICENSE file of this repository and at www.mariadb.com/bsl11.
+//
+// ANY USE OF THE LICENSED WORK IN VIOLATION OF THIS LICENSE WILL AUTOMATICALLY
+// TERMINATE YOUR RIGHTS UNDER THIS LICENSE FOR THE CURRENT AND ALL OTHER
+// VERSIONS OF THE LICENSED WORK.
+//
+// THIS LICENSE DOES NOT GRANT YOU ANY RIGHT IN ANY TRADEMARK OR LOGO OF
+// LICENSOR OR ITS AFFILIATES (PROVIDED THAT YOU MAY USE A TRADEMARK OR LOGO OF
+// LICENSOR AS EXPRESSLY REQUIRED BY THIS LICENSE).
+//
+// TO THE EXTENT PERMITTED BY APPLICABLE LAW, THE LICENSED WORK IS PROVIDED ON
+// AN “AS IS” BASIS. LICENSOR HEREBY DISCLAIMS ALL WARRANTIES AND CONDITIONS,
+// EXPRESS OR IMPLIED, INCLUDING (WITHOUT LIMITATION) WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
+// TITLE.
+
+package simulated_test
+
+import (
+	"github.com/berachain/beacon-kit/beacon/blockchain"
+	"github.com/berachain/beacon-kit/consensus/cometbft/service/encoding"
+	mathpkg "github.com/berachain/beacon-kit/primitives/math"
+	"github.com/berachain/beacon-kit/testing/simulated"
+)
+
+// TestFullLifecycle_ValidBlock_IsSuccessful tests that a valid block proposal is processed, finalized, and committed.
+// It loops through this core process `coreLoopIterations` times.
+func (s *SimulatedSuite) TestFullLifecycle_ValidBlock_IsSuccessful() {
+	const blockHeight = 1
+	const coreLoopIterations = 10
+
+	// Initialize the chain state.
+	s.initializeChain()
+
+	// Retrieve the BLS signer and proposer address.
+	blsSigner := simulated.GetBlsSigner(s.HomeDir)
+
+	// iterate through the core loop `coreLoopIterations` times, i.e. Propose, Process, Finalize and Commit.
+	proposals := s.moveChainToHeight(blockHeight, coreLoopIterations, blsSigner)
+
+	// We expect that the number of proposals that were finalized should be `coreLoopIterations`.
+	s.Require().Len(proposals, coreLoopIterations)
+
+	// Validate post-commit state.
+	queryCtx, err := s.SimComet.CreateQueryContext(blockHeight+coreLoopIterations-1, false)
+	s.Require().NoError(err)
+
+	stateDB := s.TestNode.StorageBackend.StateFromContext(queryCtx)
+	slot, err := stateDB.GetSlot()
+	s.Require().NoError(err)
+	s.Require().Equal(mathpkg.U64(blockHeight+coreLoopIterations-1), slot)
+
+	stateHeader, err := stateDB.GetLatestBlockHeader()
+	s.Require().NoError(err)
+
+	// Unmarshal the beacon block from the ABCI request.
+	proposedBlock, err := encoding.UnmarshalBeaconBlockFromABCIRequest(
+		proposals[len(proposals)-1].Txs,
+		blockchain.BeaconBlockTxIndex,
+		s.TestNode.ChainSpec.ActiveForkVersionForSlot(slot),
+	)
+	s.Require().NoError(err)
+	s.Require().Equal(proposedBlock.Message.GetHeader().GetBodyRoot(), stateHeader.GetBodyRoot())
+}
